+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.1.71](https://github.com/alternun-development/alternun/compare/v1.1.70...v1.1.71) (2026-08-30)
+
+### Changes
+
+- **repo:** 🐛 fix: complete direct source authentication ([c5b756b](https://github.com/alternun-development/alternun/commit/c5b756bf0ec457f92f32365f97bb8e232fb20d74))
+
+## [1.1.70](https://github.com/alternun-development/alternun/compare/v1.1.69...v1.1.70) (2026-08-30)
+
+### Changes
+
+- **repo:** 🐛 fix: preserve admin oidc context for google ([2a554ec](https://github.com/alternun-development/alternun/commit/2a554ec9587507969056eb22cf16462d830aaebf))
+
 ## [1.1.69](https://github.com/alternun-development/alternun/compare/v1.1.68...v1.1.69) (2026-08-30)
 
 ### Changes